Events types mysql and coreside cleanup.

+++ src/server/game/Events/GameEventMgr.cpp
@@ -190,11 +190,11 @@ void GameEventMgr::StopEvent(uint16 event_id, bool overwrite)
 
             SQLTransaction trans = CharacterDatabase.BeginTransaction();
             PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_DEL_ALL_GAME_EVENT_CONDITION_SAVE);
-            stmt->setUInt8(0, uint8(event_id));
+            stmt->setUInt16(0, event_id);
             trans->Append(stmt);
 
             stmt = CharacterDatabase.GetPreparedStatement(CHAR_DEL_GAME_EVENT_SAVE);
-            stmt->setUInt8(0, uint8(event_id));
+            stmt->setUInt16(0, event_id);
             trans->Append(stmt);
 
             CharacterDatabase.CommitTransaction(trans);
@@ -221,7 +221,7 @@ void GameEventMgr::LoadFromDB()
         {
             Field* fields = result->Fetch();
 
-            uint8 event_id = fields[0].GetUInt8();
+            uint16 event_id = fields[0].GetUInt16();
             if (event_id == 0)
             {
                 sLog->outErrorDb("`game_event` game event entry 0 is reserved and can't be used.");
@@ -284,7 +284,7 @@ void GameEventMgr::LoadFromDB()
             {
                 Field* fields = result->Fetch();
 
-                uint8 event_id = fields[0].GetUInt8();
+                uint16 event_id = fields[0].GetUInt16();
 
                 if (event_id >= mGameEvent.size())
                 {
@@ -330,7 +330,7 @@ void GameEventMgr::LoadFromDB()
             {
                 Field* fields = result->Fetch();
 
-                uint16 event_id = fields[0].GetUInt8();
+                uint16 event_id = fields[0].GetUInt16();
 
                 if (event_id >= mGameEvent.size())
                 {
@@ -384,7 +384,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 uint32 guid    = fields[0].GetUInt32();
-                int16 event_id = fields[1].GetInt8();
+                int16 event_id = fields[1].GetInt16();
 
                 int32 internal_event_id = mGameEvent.size() + event_id - 1;
 
@@ -427,7 +427,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 uint32 guid    = fields[0].GetUInt32();
-                int16 event_id = fields[1].GetInt8();
+                int16 event_id = fields[1].GetInt16();
 
                 int32 internal_event_id = mGameEvent.size() + event_id - 1;
 
@@ -471,7 +471,7 @@ void GameEventMgr::LoadFromDB()
 
                 uint32 guid     = fields[0].GetUInt32();
                 uint32 entry    = fields[1].GetUInt32();
-                uint16 event_id = fields[2].GetUInt8();
+                uint16 event_id = fields[2].GetUInt16();
 
                 if (event_id >= mGameEventModelEquip.size())
                 {
@@ -529,7 +529,7 @@ void GameEventMgr::LoadFromDB()
 
                 uint32 id       = fields[0].GetUInt32();
                 uint32 quest    = fields[1].GetUInt32();
-                uint16 event_id = fields[2].GetUInt8();
+                uint16 event_id = fields[2].GetUInt16();
 
                 if (event_id >= mGameEventCreatureQuests.size())
                 {
@@ -570,7 +570,7 @@ void GameEventMgr::LoadFromDB()
 
                 uint32 id       = fields[0].GetUInt32();
                 uint32 quest    = fields[1].GetUInt32();
-                uint16 event_id = fields[2].GetUInt8();
+                uint16 event_id = fields[2].GetUInt16();
 
                 if (event_id >= mGameEventGameObjectQuests.size())
                 {
@@ -610,7 +610,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 uint32 quest     = fields[0].GetUInt32();
-                uint16 event_id  = fields[1].GetUInt8();
+                uint16 event_id  = fields[1].GetUInt16();
                 uint32 condition = fields[2].GetUInt32();
                 float num       = fields[3].GetFloat();
 
@@ -652,7 +652,7 @@ void GameEventMgr::LoadFromDB()
             {
                 Field* fields = result->Fetch();
 
-                uint16 event_id  = fields[0].GetUInt8();
+                uint16 event_id  = fields[0].GetUInt16();
                 uint32 condition = fields[1].GetUInt32();
 
                 if (event_id >= mGameEvent.size())
@@ -694,7 +694,7 @@ void GameEventMgr::LoadFromDB()
             {
                 Field* fields = result->Fetch();
 
-                uint16 event_id  = fields[0].GetUInt8();
+                uint16 event_id  = fields[0].GetUInt16();
                 uint32 condition = fields[1].GetUInt32();
 
                 if (event_id >= mGameEvent.size())
@@ -743,7 +743,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 uint32 guid     = fields[0].GetUInt32();
-                uint16 event_id = fields[1].GetUInt8();
+                uint16 event_id = fields[1].GetUInt16();
                 uint32 npcflag  = fields[2].GetUInt32();
 
                 if (event_id >= mGameEvent.size())
@@ -783,7 +783,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 uint32 questId  = fields[0].GetUInt32();
-                uint32 eventEntry = fields[1].GetUInt32(); // TODO: Change to uint8
+                uint32 eventEntry = fields[1].GetUInt16();
 
                 Quest* questTemplate = const_cast<Quest*>(sObjectMgr->GetQuestTemplate(questId));
 
@@ -799,7 +799,7 @@ void GameEventMgr::LoadFromDB()
                     continue;
                 }
 
-                questTemplate->SetEventIdForQuest((uint16)eventEntry);
+                questTemplate->SetEventIdForQuest(eventEntry);
                 ++count;
             }
             while (result->NextRow());
@@ -828,7 +828,7 @@ void GameEventMgr::LoadFromDB()
             {
                 Field* fields = result->Fetch();
 
-                uint8 event_id  = fields[0].GetUInt8();
+                uint16 event_id  = fields[0].GetUInt16();
 
                 if (event_id >= mGameEventVendors.size())
                 {
@@ -894,7 +894,7 @@ void GameEventMgr::LoadFromDB()
             {
                 Field* fields = result->Fetch();
 
-                uint16 event_id = fields[0].GetUInt8();
+                uint16 event_id = fields[0].GetUInt16();
 
                 if (event_id >= mGameEvent.size())
                 {
@@ -934,7 +934,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 uint32 entry   = fields[0].GetUInt32();
-                int16 event_id = fields[1].GetInt8();
+                int16 event_id = fields[1].GetInt16();
 
                 int32 internal_event_id = mGameEvent.size() + event_id - 1;
 
@@ -987,7 +987,7 @@ void GameEventMgr::Initialize()
     {
         Field* fields = result->Fetch();
 
-        uint32 maxEventId = fields[0].GetUInt8();
+        uint16 maxEventId = fields[0].GetUInt16();
 
         // Id starts with 1 and vector with 0, thus increment
         maxEventId++;
@@ -1025,7 +1025,7 @@ void GameEventMgr::StartArenaSeason()
     }
 
     Field* fields = result->Fetch();
-    uint16 eventId = fields[0].GetUInt8();
+    uint16 eventId = fields[0].GetUInt16();
 
     if (eventId >= mGameEvent.size())
     {
@@ -1586,12 +1586,12 @@ void GameEventMgr::HandleQuestComplete(uint32 quest_id)
                 SQLTransaction trans = CharacterDatabase.BeginTransaction();
 
                 PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_DEL_GAME_EVENT_CONDITION_SAVE);
-                stmt->setUInt8(0, uint8(event_id));
+                stmt->setUInt16(0, event_id);
                 stmt->setUInt32(1, condition);
                 trans->Append(stmt);
 
                 stmt = CharacterDatabase.GetPreparedStatement(CHAR_INS_GAME_EVENT_CONDITION_SAVE);
-                stmt->setUInt8(0, uint8(event_id));
+                stmt->setUInt16(0, event_id);
                 stmt->setUInt32(1, condition);
                 stmt->setFloat(2, citr->second.done);
                 trans->Append(stmt);
@@ -1631,11 +1631,11 @@ void GameEventMgr::SaveWorldEventStateToDB(uint16 event_id)
     SQLTransaction trans = CharacterDatabase.BeginTransaction();
 
     PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_DEL_GAME_EVENT_SAVE);
-    stmt->setUInt8(0, uint8(event_id));
+    stmt->setUInt16(0, event_id);
     trans->Append(stmt);
 
     stmt = CharacterDatabase.GetPreparedStatement(CHAR_INS_GAME_EVENT_SAVE);
-    stmt->setUInt8(0, uint8(event_id));
+    stmt->setUInt16(0, event_id);
     stmt->setUInt8(1, mGameEvent[event_id].state);
     stmt->setUInt32(2, mGameEvent[event_id].nextstart ? uint32(mGameEvent[event_id].nextstart) : 0);
     trans->Append(stmt);
